Add files via upload
Network Sniffer - Task 1

Description:
-------------
This Python script captures network packets from the network interface in real-time using the Scapy library.
It analyzes each captured packet to extract useful information such as the source IP address, destination IP address,
and the protocol used. The purpose of this sniffer is to help understand the flow of data packets across the network and
to gain familiarity with common network protocols.